Fixed ReSharper warnings
